initial api.client refactor; move json classes into client
