+++ API.Client/API.Client.csproj
@@ -8,6 +8,11 @@
     <ItemGroup>
       <ProjectReference Include="..\DLCS.Core\DLCS.Core.csproj" />
       <ProjectReference Include="..\DLCS.Model\DLCS.Model.csproj" />
+      <ProjectReference Include="..\DLCS.Web\DLCS.Web.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
     </ItemGroup>
 
 </Project>
