+++ API.Client/JsonLd/PartialCollectionView.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace API.JsonLd
+namespace API.Client.JsonLd
 {
     public class PartialCollectionView : JsonLdBase
     {
