+++ API/Features/Image/ImageController.cs
@@ -2,8 +2,8 @@
 using System.Net;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using API.Client.JsonLd;
 using API.Features.Image.Commands;
-using API.Features.Image.Models;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
