+++ Portal.Tests/ClaimsPrincipalXTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Security.Claims;
+using API.Client;
 using FluentAssertions;
 using Xunit;
