+++ Portal/Features/Images/Commands/IngestSingleImage.cs
@@ -2,14 +2,14 @@
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
-using API.Features.Image.Models;
+using API.Client;
+using API.Client.JsonLd;
 using DLCS.Core.Settings;
 using DLCS.Model.Storage;
 using DLCS.Repository.Spaces;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Portal.Legacy;
 
 namespace Portal.Features.Images.Commands
 {
