+++ Portal/Features/Keys/Commands/CreateNewApiKey.cs
@@ -1,8 +1,8 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using API.JsonLd;
+using API.Client;
+using API.Client.JsonLd;
 using MediatR;
-using Portal.Legacy;
 
 namespace Portal.Features.Keys.Commands
 {
