+++ Portal/Features/Spaces/Requests/GetAllSpaces.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
+using API.Client;
 using DLCS.Repository;
 using DLCS.Repository.Entities;
 using MediatR;
