+++ Portal/Features/Spaces/Requests/GetImage.cs
@@ -1,9 +1,9 @@
 using System.Threading;
 using System.Threading.Tasks;
-using API.JsonLd;
+using API.Client;
+using API.Client.JsonLd;
 using MediatR;
 using Portal.Features.Spaces.Models;
-using Portal.Legacy;
 
 namespace Portal.Features.Spaces.Requests
 {
