+++ Portal/Pages/Spaces/Details.cshtml
@@ -1,5 +1,4 @@
 ﻿@page "{id}"
-@using API.JsonLd
 @model Portal.Pages.Spaces.Details
 
 @{
