+++ Portal/Portal.csproj
@@ -57,4 +57,8 @@
       <_ContentIncludedByDefault Remove="wwwroot\lib\bootstrap\LICENSE" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Legacy" />
+    </ItemGroup>
+
 </Project>
